IE 9?/10?/11 bughandling
Personally invite IE to scroll at the same rate as other browsers and
not 120x slower. No, really: why does IE divide the scroll distance by
120 and then expect the input not to be divided by 120?
Force IE to refresh the canvas on window resize. As far as I can tell,
the only way to do this is to reload the page.